override splitting of field names passed to get_field2
There are cases where we want to prevent splitting of field names that
contain dot.  For example, when this is used in
check_hadoop_namenode_state.pl, Cloudera uses the field tag.HAState in
the jmx and we don’t want to split on the dot, otherwise the field
contents never get found.