added function curl_redirect
In the case of HA + Kerberos, one faces the problem that a redirect sends one to a host for which one has no suitable certificate. However, a redirect shows that at least the passive side is working (by issuing the redirect), and that is what should be reported by Nagios. This functionality is provided by this new utility function.